requirement.txt
absl-py==0.12.0
appdirs==1.4.4
asgiref==3.3.4
asn1crypto==0.24.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Automat==0.6.0
bleach==3.3.0
blinker==1.4
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==6.7
cloud-init==21.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.2
distro-info===0.18ubuntu0.18.04.1
Django==3.2.2
ec2-hibinit-agent==1.0.0
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
gast==0.2.2
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
hibagent==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
imageio==2.9.0
importlib-metadata==4.4.0
importlib-resources==5.1.4
incremental==16.10.1
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
Jinja2==2.11.3
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
mysqlclient==2.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
PAM==0.4.2
pandocfilters==1.4.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
portpicker==1.2.0
prompt-toolkit==1.0.18
protobuf==3.16.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
Pygments==2.9.0
pygobject==3.26.1
PyJWT==1.5.3
PyMySQL==1.0.2
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.4
python-apt==1.6.5+ubuntu0.6
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.7.2
scipy==1.1.0
SecretStorage==2.3.1
service-identity==16.0.0
simplegeneric==0.8.1
six==1.15.0
sos==4.1
sqlparse==0.4.1
ssh-import-id==5.7
systemd-python==234
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
testpath==0.4.4
tornado==4.5.3
traitlets==4.3.3
Twisted==17.9.0
typing-extensions==3.7.4.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.4
virtualenv==20.4.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
zope.interface==4.3.2
